ethan sanders doesnt know what a double is

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/core/robot/tools/BetterDistanceSensor.java
@@ -47,11 +47,11 @@ public class BetterDistanceSensor extends Rev2mDistanceSensorEx {
         return this.distance;
     }
 
-    public void setRequestRate(int requestRate) {
+    public void setRequestRate(double requestRate) {
         this.requestRate = requestRate;
     }
 
-    public int getRequestRate() {
+    public double getRequestRate() {
         return this.requestRate;
     }
